add ViewHelper to examples
